+++ examples/colors.rb
@@ -1,36 +1,35 @@
 require 'rubygems'
-require 'color'
 require 'interpolate'
-
-# we need to implement +interpolate+ for Color::RGB
-# in order for Interpolation to work properly
-#
-# fortunately, Color::RGB includes +mix_with+, which is almost
-# functionally identical to +interpolate+
-class Color::RGB
-  def interpolate(other, balance)
-    mix_with(other, balance * 100.0)
-  end
-end
+require 'color'
 
 # a nice weathermap-style color gradient
 points = {
-  1 => Color::RGB::White,
+  1 => Color::RGB::Cyan,
   2 => Color::RGB::Lime,
 # 3 => ? (between Lime and Yellow; Interpolate will figure it out)
   4 => Color::RGB::Yellow,
   5 => Color::RGB::Orange,
   6 => Color::RGB::Red,
-  7 => Color::RGB::Magenta
+  7 => Color::RGB::Magenta,
+  8 => Color::RGB::White,
 }
 
-gradient = Interpolation.new(points)
+# we need to implement a blending function in order for Interpolate::Points to
+# work properly
+#
+# fortunately, Color::RGB includes +mix_with+, which is almost functionally
+# identical to what we need
 
-# what are the colors of the gradient from 1 to 7
+gradient = Interpolate::Points.new(points)
+gradient.blend_with {|color, other, balance|
+  color.mix_with(other, balance * 100.0)
+}
+
+# what are the colors of the gradient from 1 to 8
 # in increments of 0.2?
 (1).step(7, 0.2) do |value|
   color = gradient.at(value)
-  puts "A value of #{value} means #{color.html}"
+  puts "A value of #{value.round(3)} means #{color.html}"
 end